Add 2-axes coupling for angular joint limits
